feat: Minor chain improvements/fixes
