+++ src/types/index.ts
@@ -1,5 +1,6 @@
 // @index('./*.ts', f => `export * from '${f.path}'`)
 export * from './ChainAsset'
+export * from './SubstrateChain'
 export * from './SubstrateDeployment'
 export * from './SubstrateWallet'
 export * from './UseInkathonProviderContext'
